refactor(convex): centralize gear tree grouping and sort in gearTree
Extract mapSortedGearTree so public getPublishedGear and admin toPayload
share one implementation for category/item ordering.
